refactor: changed code on build pipe
