+++ pipeline/archive/name.go
@@ -1,4 +1,4 @@
-package build
+package archive
 
 import (
 	"bytes"
